+++ src/generators/errorBoundary.ts
@@ -2,9 +2,11 @@ import { WorkspaceConfiguration } from "vscode";
 
 export const generateErrorBoundary = (config: WorkspaceConfiguration) => {
   const customErrorBoundary = config.get("errorBoundary");
+
   if (customErrorBoundary) {
     return customErrorBoundary + "\n";
   }
+
   return [
     `export function ErrorBoundary(){`,
     `  const error = useRouteError();`,